Made layout more "Mobile" friendly.
Also made a drop down menu for the API's

Added a render test page. (Due to browser restrictions you can not use links to load other views. Mostly due to fact you must run an http server to use some of the api that is used to load the pages dynamically.)s